+++ internal/archive/tar_builder.go
@@ -10,16 +10,20 @@ import (
 	"time"
 
 	"github.com/pkg/errors"
+	"github.com/rafaeljusto/toglacier/internal/log"
 )
 
 // TARBuilder join all paths into an archive using the TAR computer software
 // utility.
 type TARBuilder struct {
+	logger log.Logger
 }
 
 // NewTARBuilder returns a TARBuilder with all necessary initializations.
-func NewTARBuilder() *TARBuilder {
-	return &TARBuilder{}
+func NewTARBuilder(logger log.Logger) *TARBuilder {
+	return &TARBuilder{
+		logger: logger,
+	}
 }
 
 // Build builds a tarball containing all the desired files that you want to
@@ -42,6 +46,8 @@ func NewTARBuilder() *TARBuilder {
 //       }
 //     }
 func (t TARBuilder) Build(backupPaths ...string) (string, error) {
+	t.logger.Debug("archive: creating tar file in temporary directory")
+
 	tarFile, err := ioutil.TempFile("", "toglacier-")
 	if err != nil {
 		return "", errors.WithStack(newError("", ErrorCodeTARCreation, err))
@@ -53,10 +59,12 @@ func (t TARBuilder) Build(backupPaths ...string) (string, error) {
 
 	for _, path := range backupPaths {
 		if path == "" {
+			t.logger.Info("archive: empty backup path ignored")
 			continue
 		}
 
-		if err := build(tarArchive, basePath, path); err != nil {
+		t.logger.Debugf("archive: analyzing backup path “%s”", path)
+		if err := t.build(tarArchive, basePath, path); err != nil {
 			return "", errors.WithStack(err)
 		}
 	}
@@ -65,14 +73,17 @@ func (t TARBuilder) Build(backupPaths ...string) (string, error) {
 		return "", errors.WithStack(newError(tarFile.Name(), ErrorCodeTARGeneration, err))
 	}
 
+	t.logger.Infof("archive: tar file “%s” created successfully", tarFile.Name())
 	return tarFile.Name(), nil
 }
 
-func build(tarArchive *tar.Writer, baseDir, source string) error {
+func (t TARBuilder) build(tarArchive *tar.Writer, baseDir, source string) error {
 	return filepath.Walk(source, func(path string, info os.FileInfo, err error) error {
 		if err != nil {
 			return errors.WithStack(newPathError(path, PathErrorCodeInfo, err))
 		}
+
+		t.logger.Debugf("archive: walking into path “%s”", path)
 
 		header, err := tar.FileInfoHeader(info, path)
 		if err != nil {
@@ -93,15 +104,19 @@ func build(tarArchive *tar.Writer, baseDir, source string) error {
 			header.Name += "/"
 		}
 
+		t.logger.Debugf("archive: writing tar header “%s”", header.Name)
+
 		if err = tarArchive.WriteHeader(header); err != nil {
 			return errors.WithStack(newPathError(path, PathErrorCodeWritingTARHeader, err))
 		}
 
 		if info.IsDir() {
+			t.logger.Debugf("archive: path “%s” is a directory", path)
 			return nil
 		}
 
 		if header.Typeflag != tar.TypeReg {
+			t.logger.Debugf("archive: path “%s”, with type “%d”, is not going to be added to the tar", path, header.Typeflag)
 			return nil
 		}
 
@@ -111,11 +126,12 @@ func build(tarArchive *tar.Writer, baseDir, source string) error {
 		}
 		defer file.Close()
 
-		_, err = io.CopyN(tarArchive, file, info.Size())
+		written, err := io.CopyN(tarArchive, file, info.Size())
 		if err != nil && err != io.EOF {
 			return errors.WithStack(newPathError(path, PathErrorCodeWritingFile, err))
 		}
 
+		t.logger.Debugf("archive: path “%s” copied to tar (%d bytes)", path, written)
 		return nil
 	})
 }