+++ internal/config/config_test.go
@@ -24,10 +24,13 @@ func TestDefault(t *testing.T) {
 	}{
 		{
 			description: "it should set the default configuration values",
-			expected: &config.Config{
-				AuditFile:   path.Join("var", "log", "toglacier", "audit.log"),
-				KeepBackups: 10,
-			},
+			expected: func() *config.Config {
+				c := new(config.Config)
+				c.AuditFile = path.Join("var", "log", "toglacier", "audit.log")
+				c.KeepBackups = 10
+				c.Log.Level = config.LogLevelError
+				return c
+			}(),
 		},
 	}
 
@@ -70,7 +73,9 @@ paths:
   - /usr/local/important-files-1
   - /usr/local/important-files-2
 audit file: /var/log/toglacier/audit.log
-log file: /var/log/toglacier/toglacier.log
+log:
+  file: /var/log/toglacier/toglacier.log
+  level:   DEBUG
 keep backups: 10
 backup secret: encrypted:M5rNhMpetktcTEOSuF25mYNn97TN1w==
 email:
@@ -99,7 +104,8 @@ aws:
 					"/usr/local/important-files-2",
 				}
 				c.AuditFile = "/var/log/toglacier/audit.log"
-				c.LogFile = "/var/log/toglacier/toglacier.log"
+				c.Log.File = "/var/log/toglacier/toglacier.log"
+				c.Log.Level = config.LogLevelDebug
 				c.KeepBackups = 10
 				c.BackupSecret.Value = "abc12300000000000000000000000000"
 				c.Email.Server = "smtp.example.com"
@@ -140,6 +146,53 @@ aws:
 			defer f.Close()
 
 			f.WriteString(`
+paths:
+  - /usr/local/important-files-1
+  - /usr/local/important-files-2
+audit file: /var/log/toglacier/audit.log
+log:
+  file: /var/log/toglacier/toglacier.log
+  level: idontexist
+keep backups: 10
+backup secret: encrypted:M5rNhMpetktcTEOSuF25mYNn97TN1w==
+email:
+  server: smtp.example.com
+  port: 587
+  username: user@example.com
+  password: encrypted:i9dw0HZPOzNiFgtEtrr0tiY0W+YYlA==
+  from: user@example.com
+  to:
+    - report1@example.com
+    - report2@example.com
+aws:
+  account id: encrypted:DueEGILYe8OoEp49Qt7Gymms2sPuk5weSPiG6w==
+  access key id: encrypted:XesW4TPKzT3Cgw1SCXeMB9Pb2TssRPCdM4mrPwlf4zWpzSZQ
+  secret access key: encrypted:hHHZXW+Uuj+efOA7NR4QDAZh6tzLqoHFaUHkg/Yw1GE/3sJBi+4cn81LhR8OSVhNwv1rI6BR4fA=
+  region: us-east-1
+  vault name: backup
+`)
+
+			var s scenario
+			s.description = "it should detect when the log level is unknown"
+			s.filename = f.Name()
+			s.expectedError = &config.Error{
+				Filename: f.Name(),
+				Code:     config.ErrorCodeParsingYAML,
+				Err: &config.Error{
+					Code: config.ErrorCodeLogLevel,
+				},
+			}
+
+			return s
+		}(),
+		func() scenario {
+			f, err := ioutil.TempFile("", "toglacier-")
+			if err != nil {
+				t.Fatalf("error creating a temporary file. details %s", err)
+			}
+			defer f.Close()
+
+			f.WriteString(`
 - /usr/local/important-files-1
 - /usr/local/important-files-2
 `)
@@ -171,7 +224,9 @@ paths:
   - /usr/local/important-files-1
   - /usr/local/important-files-2
 audit file: /var/log/toglacier/audit.log
-log file: /var/log/toglacier/toglacier.log
+log:
+  file: /var/log/toglacier/toglacier.log
+  level: debug
 keep backups: 10
 backup secret: encrypted:M5rNhMpetktcTEOSuF25mYNn97TN1w==
 email:
@@ -217,7 +272,9 @@ paths:
   - /usr/local/important-files-1
   - /usr/local/important-files-2
 audit file: /var/log/toglacier/audit.log
-log file: /var/log/toglacier/toglacier.log
+log:
+  file: /var/log/toglacier/toglacier.log
+  level: debug
 keep backups: 10
 backup secret: encrypted:invalid
 email:
@@ -265,7 +322,9 @@ paths:
   - /usr/local/important-files-1
   - /usr/local/important-files-2
 audit file: /var/log/toglacier/audit.log
-log file: /var/log/toglacier/toglacier.log
+log:
+  file: /var/log/toglacier/toglacier.log
+  level: debug
 keep backups: 10
 backup secret: a123456789012345678901234567890
 email:
@@ -294,7 +353,8 @@ aws:
 					"/usr/local/important-files-2",
 				}
 				c.AuditFile = "/var/log/toglacier/audit.log"
-				c.LogFile = "/var/log/toglacier/toglacier.log"
+				c.Log.File = "/var/log/toglacier/toglacier.log"
+				c.Log.Level = config.LogLevelDebug
 				c.KeepBackups = 10
 				c.BackupSecret.Value = "a1234567890123456789012345678900"
 				c.Email.Server = "smtp.example.com"
@@ -328,7 +388,9 @@ paths:
   - /usr/local/important-files-1
   - /usr/local/important-files-2
 audit file: /var/log/toglacier/audit.log
-log file: /var/log/toglacier/toglacier.log
+log:
+  file: /var/log/toglacier/toglacier.log
+  level: debug
 keep backups: 10
 backup secret: a12345678901234567890123456789012
 email:
@@ -357,7 +419,8 @@ aws:
 					"/usr/local/important-files-2",
 				}
 				c.AuditFile = "/var/log/toglacier/audit.log"
-				c.LogFile = "/var/log/toglacier/toglacier.log"
+				c.Log.File = "/var/log/toglacier/toglacier.log"
+				c.Log.Level = config.LogLevelDebug
 				c.KeepBackups = 10
 				c.BackupSecret.Value = "a1234567890123456789012345678901"
 				c.Email.Server = "smtp.example.com"
@@ -424,6 +487,7 @@ func TestLoadFromEnvironment(t *testing.T) {
 				"TOGLACIER_PATHS":                 "/usr/local/important-files-1,/usr/local/important-files-2",
 				"TOGLACIER_AUDIT":                 "/var/log/toglacier/audit.log",
 				"TOGLACIER_LOG_FILE":              "/var/log/toglacier/toglacier.log",
+				"TOGLACIER_LOG_LEVEL":             "  DEBUG  ",
 				"TOGLACIER_KEEP_BACKUPS":          "10",
 				"TOGLACIER_BACKUP_SECRET":         "encrypted:M5rNhMpetktcTEOSuF25mYNn97TN1w==",
 			},
@@ -434,7 +498,8 @@ func TestLoadFromEnvironment(t *testing.T) {
 					"/usr/local/important-files-2",
 				}
 				c.AuditFile = "/var/log/toglacier/audit.log"
-				c.LogFile = "/var/log/toglacier/toglacier.log"
+				c.Log.File = "/var/log/toglacier/toglacier.log"
+				c.Log.Level = config.LogLevelDebug
 				c.KeepBackups = 10
 				c.BackupSecret.Value = "abc12300000000000000000000000000"
 				c.Email.Server = "smtp.example.com"
@@ -455,6 +520,40 @@ func TestLoadFromEnvironment(t *testing.T) {
 			}(),
 		},
 		{
+			description: "it should detect an invalid log level",
+			env: map[string]string{
+				"TOGLACIER_AWS_ACCOUNT_ID":        "encrypted:DueEGILYe8OoEp49Qt7Gymms2sPuk5weSPiG6w==",
+				"TOGLACIER_AWS_ACCESS_KEY_ID":     "encrypted:XesW4TPKzT3Cgw1SCXeMB9Pb2TssRPCdM4mrPwlf4zWpzSZQ",
+				"TOGLACIER_AWS_SECRET_ACCESS_KEY": "encrypted:hHHZXW+Uuj+efOA7NR4QDAZh6tzLqoHFaUHkg/Yw1GE/3sJBi+4cn81LhR8OSVhNwv1rI6BR4fA=",
+				"TOGLACIER_AWS_REGION":            "us-east-1",
+				"TOGLACIER_AWS_VAULT_NAME":        "backup",
+				"TOGLACIER_EMAIL_SERVER":          "smtp.example.com",
+				"TOGLACIER_EMAIL_PORT":            "587",
+				"TOGLACIER_EMAIL_USERNAME":        "user@example.com",
+				"TOGLACIER_EMAIL_PASSWORD":        "encrypted:i9dw0HZPOzNiFgtEtrr0tiY0W+YYlA==",
+				"TOGLACIER_EMAIL_FROM":            "user@example.com",
+				"TOGLACIER_EMAIL_TO":              "report1@example.com,report2@example.com",
+				"TOGLACIER_PATHS":                 "/usr/local/important-files-1,/usr/local/important-files-2",
+				"TOGLACIER_AUDIT":                 "/var/log/toglacier/audit.log",
+				"TOGLACIER_LOG_FILE":              "/var/log/toglacier/toglacier.log",
+				"TOGLACIER_LOG_LEVEL":             "idontexist",
+				"TOGLACIER_KEEP_BACKUPS":          "10",
+				"TOGLACIER_BACKUP_SECRET":         "encrypted:M5rNhMpetktcTEOSuF25mYNn97TN1w==",
+			},
+			expectedError: &config.Error{
+				Code: config.ErrorCodeReadingEnvVars,
+				Err: &envconfig.ParseError{
+					KeyName:   "TOGLACIER_LOG_LEVEL",
+					FieldName: "Level",
+					TypeName:  "config.LogLevel",
+					Value:     "idontexist",
+					Err: &config.Error{
+						Code: config.ErrorCodeLogLevel,
+					},
+				},
+			},
+		},
+		{
 			description: "it should detect invalid encrypted values",
 			env: map[string]string{
 				"TOGLACIER_AWS_ACCOUNT_ID":        "encrypted:invalid",
@@ -471,6 +570,7 @@ func TestLoadFromEnvironment(t *testing.T) {
 				"TOGLACIER_PATHS":                 "/usr/local/important-files-1,/usr/local/important-files-2",
 				"TOGLACIER_AUDIT":                 "/var/log/toglacier/audit.log",
 				"TOGLACIER_LOG_FILE":              "/var/log/toglacier/toglacier.log",
+				"TOGLACIER_LOG_LEVEL":             "debug",
 				"TOGLACIER_KEEP_BACKUPS":          "10",
 				"TOGLACIER_BACKUP_SECRET":         "encrypted:M5rNhMpetktcTEOSuF25mYNn97TN1w==",
 			},
@@ -505,6 +605,7 @@ func TestLoadFromEnvironment(t *testing.T) {
 				"TOGLACIER_PATHS":                 "/usr/local/important-files-1,/usr/local/important-files-2",
 				"TOGLACIER_AUDIT":                 "/var/log/toglacier/audit.log",
 				"TOGLACIER_LOG_FILE":              "/var/log/toglacier/toglacier.log",
+				"TOGLACIER_LOG_LEVEL":             "debug",
 				"TOGLACIER_KEEP_BACKUPS":          "10",
 				"TOGLACIER_BACKUP_SECRET":         "encrypted:invalid",
 			},
@@ -539,6 +640,7 @@ func TestLoadFromEnvironment(t *testing.T) {
 				"TOGLACIER_PATHS":                 "/usr/local/important-files-1,/usr/local/important-files-2",
 				"TOGLACIER_AUDIT":                 "/var/log/toglacier/audit.log",
 				"TOGLACIER_LOG_FILE":              "/var/log/toglacier/toglacier.log",
+				"TOGLACIER_LOG_LEVEL":             "debug",
 				"TOGLACIER_KEEP_BACKUPS":          "10",
 				"TOGLACIER_BACKUP_SECRET":         "a123456789012345678901234567890",
 			},
@@ -549,7 +651,8 @@ func TestLoadFromEnvironment(t *testing.T) {
 					"/usr/local/important-files-2",
 				}
 				c.AuditFile = "/var/log/toglacier/audit.log"
-				c.LogFile = "/var/log/toglacier/toglacier.log"
+				c.Log.File = "/var/log/toglacier/toglacier.log"
+				c.Log.Level = config.LogLevelDebug
 				c.KeepBackups = 10
 				c.BackupSecret.Value = "a1234567890123456789012345678900"
 				c.Email.Server = "smtp.example.com"
@@ -586,6 +689,7 @@ func TestLoadFromEnvironment(t *testing.T) {
 				"TOGLACIER_PATHS":                 "/usr/local/important-files-1,/usr/local/important-files-2",
 				"TOGLACIER_AUDIT":                 "/var/log/toglacier/audit.log",
 				"TOGLACIER_LOG_FILE":              "/var/log/toglacier/toglacier.log",
+				"TOGLACIER_LOG_LEVEL":             "debug",
 				"TOGLACIER_KEEP_BACKUPS":          "10",
 				"TOGLACIER_BACKUP_SECRET":         "a12345678901234567890123456789012",
 			},
@@ -596,7 +700,8 @@ func TestLoadFromEnvironment(t *testing.T) {
 					"/usr/local/important-files-2",
 				}
 				c.AuditFile = "/var/log/toglacier/audit.log"
-				c.LogFile = "/var/log/toglacier/toglacier.log"
+				c.Log.File = "/var/log/toglacier/toglacier.log"
+				c.Log.Level = config.LogLevelDebug
 				c.KeepBackups = 10
 				c.BackupSecret.Value = "a1234567890123456789012345678901"
 				c.Email.Server = "smtp.example.com"